重新修改了项目结构,根据要求重写了代码.

+++ 实验一/实验1-2 线性表链式存储结构的操作与应用/LinkedList.c
@@ -1,15 +1,14 @@
-#include <stdint.h>
 #include <stdio.h>
 #include <stdlib.h>
 
-#define ELEMTYPE    int
-#define BOOL        int
-#define TRUE    1
-#define FALSE   0
+#define ELEMTYPE int
+#define BOOL     int
+#define TRUE     1
+#define FALSE    0
 
 typedef struct LNode {
-    ELEMTYPE     data;
-    struct LNode *next;
+    ELEMTYPE data;
+    struct LNode* next;
 } LNode, *PLNode;
 
 // 传入一个指向NULL的指针
@@ -17,7 +16,7 @@ typedef struct LNode {
 PLNode CreateListFromTail(PLNode pl) {
     PLNode temp = NULL, new_pl = NULL;
     char buffer[256];
-    
+
     // 为指针分配内存,初始化节点内容
     pl = (PLNode)malloc(sizeof(LNode));
     pl->next = NULL;
@@ -25,8 +24,8 @@ PLNode CreateListFromTail(PLNode pl) {
 
     scanf_s("%s", buffer, sizeof(buffer));
     ELEMTYPE c = strtol(buffer, NULL, 10);
-    while(c != -1) {
-        new_pl = (PLNode)malloc(sizeof(LNode)); 
+    while (c != -1) {
+        new_pl = (PLNode)malloc(sizeof(LNode));
         new_pl->data = c;
         new_pl->next = NULL;
 
@@ -44,10 +43,13 @@ PLNode CreateListFromTail(PLNode pl) {
 PLNode DeleteListElem(PLNode pl, size_t index) {
     PLNode temp = pl;
     for (size_t i = 1; i < index; temp = temp->next) {
-        if (temp->next == NULL) { puts("index 超出了范围\n"); }
+        if (temp->next == NULL) {
+            puts("index 超出了范围\n");
+        }
         i++;
     }
-    printf("你当前想删除的位置是: %d -> 对应的值是: %d\n", index, temp->next->data);
+    printf("你当前想删除的位置是: %d -> 对应的值是: %d\n", (int)index,
+           temp->next->data);
     temp->next = temp->next->next;
 
     return pl;
@@ -57,10 +59,13 @@ PLNode DeleteListElem(PLNode pl, size_t index) {
 PLNode ModifyListElem(PLNode pl, size_t index, ELEMTYPE md) {
     PLNode temp = pl;
     for (size_t i = 1; i < index; temp = temp->next) {
-        if (temp->next == NULL) { puts("index 超出了范围\n"); }
+        if (temp->next == NULL) {
+            puts("index 超出了范围\n");
+        }
         i++;
     }
-    printf("你当前想删除的位置是: %d -> 对应的值是: %d -> 修改之后的值为: %d\n", index, temp->next->data, md);
+    printf("你当前想删除的位置是: %d -> 对应的值是: %d -> 修改之后的值为: %d\n",
+           (int)index, temp->next->data, md);
     temp->next->data = md;
 
     return pl;
@@ -72,7 +77,7 @@ BOOL PrintInfor(PLNode pl) {
     int count = 1;
     for (; temp->next != NULL; temp = temp->next) {
         printf("List %-5d -> value is: %d\n", count, temp->data);
-        count++; 
+        count++;
     }
 
     return TRUE;
@@ -82,7 +87,7 @@ int main() {
     puts("请输入数值以创造队列");
     puts("用-1结束输入");
     PLNode pl;
-    pl = CreateListFromTail(pl); 
+    pl = CreateListFromTail(pl);
     PrintInfor(pl);
 
     DeleteListElem(pl, 3);
